Switch away from the state-manager, and thus FFI for non-fvm commands
