+++ ephemeralbs.go
@@ -1,3 +1,5 @@
+//go:build unix
+
 package main
 
 import (
