+++ state.go
@@ -3,7 +3,6 @@ package main
 import (
 	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
 	"sync"
 	"sync/atomic"
@@ -16,17 +15,11 @@ import (
 	filadt "github.com/filecoin-project/go-state-types/builtin/v13/util/adt"
 	filstore "github.com/filecoin-project/go-state-types/store"
 
-	"github.com/filecoin-project/lotus/build"
-	"github.com/filecoin-project/lotus/chain/actors"
-	"github.com/filecoin-project/lotus/chain/types/ethtypes"
-	carbs "github.com/ipld/go-car/v2/blockstore"
-
 	lbi "github.com/filecoin-project/lotus/chain/actors/builtin"
 	lbimsig "github.com/filecoin-project/lotus/chain/actors/builtin/multisig"
-	lchstmgr "github.com/filecoin-project/lotus/chain/stmgr"
+	lchstate "github.com/filecoin-project/lotus/chain/state"
 	lchtypes "github.com/filecoin-project/lotus/chain/types"
 
-	"github.com/ipfs/go-cid"
 	ipldcbor "github.com/ipfs/go-ipld-cbor"
 	cbg "github.com/whyrusleeping/cbor-gen"
 	"golang.org/x/xerrors"
@@ -35,13 +28,8 @@ import (
 var hamtOptions = append(filadt.DefaultHamtOptions, hamt.UseTreeBitWidth(filbuiltin.DefaultHamtBitwidth))
 
 func getCoins(ctx context.Context, bg *blockGetter, ts *lchtypes.TipSet, addr filaddr.Address) error {
-	sm, err := newFilStateReader(bg)
-	if err != nil {
-		return xerrors.Errorf("unable to initialize a StateManager: %w", err)
-	}
-
 	foundAttoFil := filabi.NewTokenAmount(0)
-	if err := parseActors(ctx, sm, ts, addr, foundAttoFil); err != nil {
+	if err := parseActors(ctx, bg, ts, addr, foundAttoFil); err != nil {
 		return err
 	}
 
@@ -50,12 +38,13 @@ func getCoins(ctx context.Context, bg *blockGetter, ts *lchtypes.TipSet, addr fi
 	return nil
 }
 
-func parseActors(ctx context.Context, sm *lchstmgr.StateManager, ts *lchtypes.TipSet, rootAddr filaddr.Address, foundAttoFil filabi.TokenAmount) error {
-	ast := filstore.WrapStore(ctx, ipldcbor.NewCborStore(sm.ChainStore().UnionStore()))
+func parseActors(ctx context.Context, bg *blockGetter, ts *lchtypes.TipSet, rootAddr filaddr.Address, foundAttoFil filabi.TokenAmount) error {
+	ast := filstore.WrapStore(ctx, ipldcbor.NewCborStore(bg))
 	getManyAst := &getManyCborStore{
-		BasicIpldStore: ipldcbor.NewCborStore(sm.ChainStore().StateBlockstore())}
+		BasicIpldStore: ipldcbor.NewCborStore(bg),
+	}
 
-	stateTree, err := sm.StateTree(ts.ParentState())
+	stateTree, err := lchstate.LoadStateTree(ipldcbor.NewCborStore(bg), ts.ParentState())
 	if err != nil {
 		return err
 	}
@@ -134,16 +123,12 @@ func parseActors(ctx context.Context, sm *lchstmgr.StateManager, ts *lchtypes.Ti
 }
 
 func getActors(ctx context.Context, bg *blockGetter, ts *lchtypes.TipSet, countOnly bool) error {
-	sm, err := newFilStateReader(bg)
-	if err != nil {
-		return xerrors.Errorf("unable to initialize a StateManager: %w", err)
-	}
-
 	var numActors uint64
 
-	ast := filstore.WrapStore(ctx, ipldcbor.NewCborStore(sm.ChainStore().UnionStore()))
+	ast := filstore.WrapStore(ctx, ipldcbor.NewCborStore(bg))
 	getManyAst := &getManyCborStore{
-		BasicIpldStore: ipldcbor.NewCborStore(sm.ChainStore().StateBlockstore())}
+		BasicIpldStore: ipldcbor.NewCborStore(bg),
+	}
 
 	var root lchtypes.StateRoot
 	// Try loading as a new-style state-tree (version/actors tuple).
@@ -186,12 +171,7 @@ func getActors(ctx context.Context, bg *blockGetter, ts *lchtypes.TipSet, countO
 }
 
 func getBalance(ctx context.Context, bg *blockGetter, ts *lchtypes.TipSet, addr filaddr.Address) error {
-	sm, err := newFilStateReader(bg)
-	if err != nil {
-		return xerrors.Errorf("unable to initialize a StateManager: %w", err)
-	}
-
-	stateTree, err := sm.StateTree(ts.ParentState())
+	stateTree, err := lchstate.LoadStateTree(ipldcbor.NewCborStore(bg), ts.ParentState())
 	if err != nil {
 		return err
 	}
@@ -203,155 +183,4 @@ func getBalance(ctx context.Context, bg *blockGetter, ts *lchtypes.TipSet, addr 
 	fmt.Printf("total actor balance: %s\n", act.Balance)
 	bg.PrintStats()
 	return nil
-}
-
-func fevmExec(ctx context.Context, bg *blockGetter, ts *lchtypes.TipSet, eaddr *ethtypes.EthAddress, edata ethtypes.EthBytes, outputCAR string) error {
-	filMsg, err := getFevmRequest(eaddr, edata)
-	if err != nil {
-		return err
-	}
-	fmt.Printf("epoch %s\n", ts.Height())
-
-	sm, err := newFilStateReader(bg)
-	if err != nil {
-		return err
-	}
-
-	act, err := sm.LoadActor(ctx, filMsg.To, ts)
-	if err != nil {
-		return xerrors.Errorf("could not load actor the message is from: %w", err)
-	}
-	actorStateRoot := act.Head
-	fmt.Printf("actor state root: %s\n", actorStateRoot)
-
-	_, err = bg.Get(ctx, actorStateRoot)
-	if err != nil {
-		return fmt.Errorf("error loading state root %w", err)
-	}
-
-	for _, c := range bg.orderedCids {
-		fmt.Printf("pre-call cid: %s\n", c)
-	}
-
-	res, err := sm.Call(ctx, filMsg, ts)
-	if err != nil {
-		return xerrors.Errorf("unable to make a call: %w", err)
-	}
-
-	str, err := json.Marshal(res)
-	if err != nil {
-		fmt.Println(res)
-		return xerrors.Errorf("could not marshal response as json: %w", err)
-	}
-
-	fmt.Println(string(str))
-
-	bg.PrintStats()
-
-	cidsInOrder := bg.orderedCids
-	if err != nil {
-		return err
-	}
-
-	for _, c := range cidsInOrder {
-		fmt.Printf("all cid: %s\n", c)
-	}
-
-	if outputCAR != "" {
-		carw, err := carbs.OpenReadWrite(outputCAR, []cid.Cid{actorStateRoot}, carbs.WriteAsCarV1(true))
-		if err != nil {
-			return err
-		}
-
-		for _, c := range cidsInOrder {
-			blk, err := bg.Get(ctx, c)
-			if err != nil {
-				return err
-			}
-			if err := carw.Put(ctx, blk); err != nil {
-				return err
-			}
-		}
-
-		if err := carw.Finalize(); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func getFevmRequest(eaddr *ethtypes.EthAddress, edata ethtypes.EthBytes) (*lchtypes.Message, error) {
-	tx := ethtypes.EthCall{
-		From:     nil,
-		To:       eaddr,
-		Gas:      0,
-		GasPrice: ethtypes.EthBigInt{},
-		Value:    ethtypes.EthBigInt{},
-		Data:     edata,
-	}
-	filMsg, err := ethCallToFilecoinMessage(tx)
-	if err != nil {
-		return nil, xerrors.Errorf("unable to convert ethcall to filecoin message: %w", err)
-	}
-
-	return filMsg, nil
-}
-
-func ethCallToFilecoinMessage(tx ethtypes.EthCall) (*lchtypes.Message, error) {
-	var from filaddr.Address
-	if tx.From == nil || *tx.From == (ethtypes.EthAddress{}) {
-		// Send from the filecoin "system" address.
-		var err error
-		from, err = (ethtypes.EthAddress{}).ToFilecoinAddress()
-		if err != nil {
-			return nil, fmt.Errorf("failed to construct the ethereum system address: %w", err)
-		}
-	} else {
-		// The from address must be translatable to an f4 address.
-		var err error
-		from, err = tx.From.ToFilecoinAddress()
-		if err != nil {
-			return nil, fmt.Errorf("failed to translate sender address (%s): %w", tx.From.String(), err)
-		}
-		if p := from.Protocol(); p != filaddr.Delegated {
-			return nil, fmt.Errorf("expected a class 4 address, got: %d: %w", p, err)
-		}
-	}
-
-	var params []byte
-	if len(tx.Data) > 0 {
-		initcode := filabi.CborBytes(tx.Data)
-		params2, err := actors.SerializeParams(&initcode)
-		if err != nil {
-			return nil, fmt.Errorf("failed to serialize params: %w", err)
-		}
-		params = params2
-	}
-
-	var to filaddr.Address
-	var method filabi.MethodNum
-	if tx.To == nil {
-		// this is a contract creation
-		to = filbuiltin.EthereumAddressManagerActorAddr
-		method = filbuiltin.MethodsEAM.CreateExternal
-	} else {
-		addr, err := tx.To.ToFilecoinAddress()
-		if err != nil {
-			return nil, xerrors.Errorf("cannot get Filecoin address: %w", err)
-		}
-		to = addr
-		method = filbuiltin.MethodsEVM.InvokeContract
-	}
-
-	return &lchtypes.Message{
-		From:       from,
-		To:         to,
-		Value:      filbig.Int(tx.Value),
-		Method:     method,
-		Params:     params,
-		GasLimit:   build.BlockGasLimit,
-		GasFeeCap:  filbig.Zero(),
-		GasPremium: filbig.Zero(),
-	}, nil
 }
